move client.py to dbus/
